+++ src/drawing/mod.rs
@@ -6,6 +6,6 @@ pub(crate) mod layer;
 pub(crate) mod scene;
 pub use layer::draw_layer;
 pub use scene::{
-    draw_scene, node_tree_list, node_tree_list_visible, paint_node_tree, print_scene,
-    render_node_tree, vibrancy_color_filter, ExternalBackdrop,
+    draw_scene, node_surfaces_stats, node_tree_list, node_tree_list_visible, paint_node_tree,
+    print_scene, render_node_tree, vibrancy_color_filter, ExternalBackdrop,
 };